fix: correct Dapr firestore component field names and daprd readiness check (#252)
- Rename projectId to project_id in statestore-local.yaml (snake_case required
  by Dapr v1.18 state.gcp.firestore component; camelCase caused fatal exit)
- Remove invalid databaseId field (not a valid Dapr component field)
- Add endpoint: "localhost:8081" for documented emulator connection
- Add --dapr-http-port 3500 for explicit health probe target
- Fix dev-up readiness loop: track probe success, dump daprd log on failure,
  exit 1 instead of false-positive "daprd ready" message
- Add gcloud components install (beta + cloud-datastore-emulator) to
  dev-dapr-setup for fresh environments
- Add --quiet flag to gcloud emulator command for non-interactive use
- Fix integration test: Players.Current is int, not int32

+++ internal/db/dapr_integration_test.go
@@ -55,7 +55,7 @@ func TestDaprDB_Integration_StatusCRUD(t *testing.T) {
 
 	got, err = db.GetStatus(ctx, instance)
 	require.NoError(t, err)
-	assert.Equal(t, int32(5), got.Players.Current)
+	assert.Equal(t, 5, got.Players.Current)
 }
 
 func TestDaprDB_Integration_WhitelistCRUD(t *testing.T) {
